feat: update user access state to align with backend

+++ frontend/src/hooks/subtitles/useRemoveAccess.ts
@@ -1,12 +1,22 @@
 import { useCallback } from "react";
 import { toast } from "react-hot-toast";
-import useSubtitleDocStore from "../../store/useSubtitleDocStore";
+import useSubtitleDocStore, {
+  SubtitleAccess,
+  AccessType,
+} from "../../store/useSubtitleDocStore";
 
 export const useRemoveAccess = () => {
   const { updateAccess, docs } = useSubtitleDocStore();
 
   const removeAccess = useCallback(
     async (docId: string, userId: string) => {
+      const doc = docs.find((doc) => doc.id === docId);
+      if (!doc) return;
+
+      const user = doc.access.find((u) => u.userId === userId);
+      if (!user) return;
+
+      const isDowngrade = user.accessType === "EDIT";
       try {
         const response = await fetch(`/api/subtitles/${docId}/access`, {
           method: "DELETE",
@@ -16,22 +26,23 @@ export const useRemoveAccess = () => {
         });
 
         const data = await response.json();
-
         if (!response.ok) {
-          throw new Error(data.message || "Failed to remove access");
+          throw new Error(data.message || "Failed to update access");
         }
 
-        const updatedDoc = docs.find((doc) => doc.id === docId);
-        if (!updatedDoc) return;
-
-        const newAccessList = updatedDoc.access.filter(
-          (u) => u.userId !== userId
-        );
-        updateAccess(docId, newAccessList);
-
         toast.success(data.message);
+
+        const updatedAccessList: SubtitleAccess[] = isDowngrade
+          ? doc.access.map((u) =>
+              u.userId === userId
+                ? { ...u, accessType: "VIEW" as AccessType }
+                : u
+            )
+          : doc.access.filter((u) => u.userId !== userId);
+
+        updateAccess(docId, updatedAccessList);
       } catch (error: any) {
-        toast.error(error.message || "Failed to remove user access.");
+        toast.error(error.message || "Failed to update access.");
       }
     },
     [updateAccess, docs]
